requirements.txt
streamlit
numpy
tensorflow
pillow
python-dotenv
google-generativeai
